C++23: the last of the size_ts?
(cherry picked from commit 13fd474ffddacbc7261e053199a0727900f9a53d)

+++ Exercises/Modules/Chapter 08/Soln8_07.cpp
@@ -1,23 +1,23 @@
 // Exercise 8-7 Computing Fibonacci numbers recursively. 
-// Main thing to realise is that the recursion needs two base cases.
+// Main thing to realize is that the recursion needs two base cases.
 // Key is also to use unsigned values (function would fail for negative numbers)
 // and not to forget about zero either (using n == 1 and n == 2 as base cases 
 // would mean trouble if n == 0 is passed)
 import std;
 
-unsigned long long fib(size_t n);
+unsigned long long fib(unsigned int n);
 
 int main()
 {
-  size_t num{};
+  unsigned int num{};
   std::print("Good day, master. How many Fibonacci numbers shall I compute today? ");
   std::cin >> num;
   
-  for (size_t i{1}; i <= num; ++i)
-     std::println("fib({}) = {}", i, fib(i));
+  for (unsigned int i{1}; i <= num; ++i)
+    std::println("fib({}) = {}", i, fib(i));
 }
 
-unsigned long long fib(size_t n)
+unsigned long long fib(unsigned int n)
 {
   switch (n)
   {
