+++ Exercises/Modules/Chapter 14/Soln14_04/Person.cpp
@@ -3,7 +3,7 @@
 module person;
 import std;
 
-Person::Person(size_t age, std::string_view name, Gender gender) 
+Person::Person(unsigned age, std::string_view name, Gender gender) 
   : m_age {age}, m_name {name}, m_gender {gender}
 {
   // Instead of just initializing the members with the argument values,
